07_djnago_paginations_search_filter/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
Faker==20.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
